Move reducer types into separate file
